feat: stage from anywhere in working tree
The index.stage() method takes an additional parameter so that it has
both the path to the file from the current working directory (i.e. the
directory from which the gitlet binary was executed) and its path
relative to the root of the working tree. The former is used to open and
read the file for hashing; the latter for entry into the index.